handle optionals during drift detection and updates for namespace
Since pulsar have some defaults, it's quite beneficial to allow
undefined values in the tf definition.
That basically means "not managed by tf". But in the previous state
of this provider null-values were submitted in any case,
sometimes silently breaking a valid setup or just facing a validation on
the pulsar side and failing an update.

This PR brings per-prop check, to avoid unnecessary api calls and be more precise on what is really updating.
That way if any of the optionals left empty, it wouldn't be changed on the pulsar side.

To mitigate some tf limitations, where tf doesn't know if props of an element of the set defined,
this PR brings some validation and assuming defaults which are invalid,
to simply ignore it later on update/create.

This leads to side effect of never empty plan after update when only part of the `namespace_config` specified.
That should encourage user to define all the properties. Another option would be to flatten this config and
use tf capabilities to check if its defined, but that would require migration of the state and config.