importing pulsar resources (#14)
Now its possible to do `terraform import pulsar_<type> <id>` to get
existing resources states. However even in the case of direct resource
creation with already existing ones it would be gracefully handled.

Also extended slightly read method for cluster and namespace fixing
an issue with incomplete read.

Also to give a feedback on very early stages added validation of
parameters, as it's valuable for import

fixes #13


The PR looks big, but actually the logic is almost the same - import method, then tests for the existing resource. But if it would be hard to review, I'm ready to split it per resource.